+++ interfaces/interface/Cargo.toml
@@ -7,5 +7,5 @@ edition = "2018"
 
 [dependencies]
 futures = { version = "0.3.1", default-features = false, features = ["alloc"] }
-redshirt-syscalls = { path = "../syscalls", default-features = false }
 parity-scale-codec = { version = "1.0.5", default-features = false, features = ["derive"] }
+redshirt-syscalls = { path = "../syscalls", default-features = false }
