+++ interfaces/syscalls/src/ffi.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{EncodedMessage, InterfaceHash, MessageId, Pid};
+use crate::{EncodedMessage, MessageId};
 
 use alloc::vec::Vec;
 use core::convert::TryFrom as _;
@@ -23,10 +23,9 @@ use core::convert::TryFrom as _;
 extern "C" {
     /// Asks for the next notification.
     ///
-    /// The `to_poll` parameter must be a list (whose length is `to_poll_len`) of notifications to
-    /// poll. Entries in this list equal to `0` are ignored. Entries equal to `1` are special and
-    /// mean "a message received on an interface or a process destroyed notification". If a
-    /// notification is successfully pulled, the corresponding entry in `to_poll` is set to `0`.
+    /// The `to_poll` parameter must be a list (whose length is `to_poll_len`) of messages whose
+    /// answer to poll. Entries in this list equal to `0` are ignored. If a notification is
+    /// successfully pulled, the corresponding entry in `to_poll` is set to `0`.
     ///
     /// Flags is a bitfield, defined as:
     ///
@@ -46,7 +45,7 @@ extern "C" {
     /// notifications for a notification that fits in `out_len`. It will however skip the
     /// notifications in the queue that do not match any entry in `to_poll`.
     ///
-    /// Messages written in `out` can be decoded into a [`DecodedNotification`].
+    /// Messages written in `out` can be decoded using [`decode_notification`].
     ///
     /// When this function is being called, a "lock" is being held on the memory pointed by
     /// `to_poll` and `out`. In particular, it is invalid to modify these buffers while the
@@ -139,239 +138,14 @@ extern "C" {
     pub(crate) fn cancel_message(message_id: *const u64);
 }
 
-/// Prototype for a message.
+// TODO: all the decoding performs unaligned reads, which isn't great
+
 #[derive(Debug, Clone)]
-pub enum NotificationBuilder {
-    /// Prototype for an interface message.
-    Interface(InterfaceNotificationBuilder),
-    /// Prototype for a response message.
-    Response(ResponseNotificationBuilder),
-    /// Prototype for a process destroyed message.
-    ProcessDestroyed(ProcessDestroyedNotificationBuilder),
+pub struct NotificationBuilder {
+    data: Vec<u8>,
 }
 
 impl NotificationBuilder {
-    /// Returns the length in bytes of the constructed message.
-    pub fn len(&self) -> usize {
-        match self {
-            NotificationBuilder::Interface(msg) => msg.len(),
-            NotificationBuilder::Response(msg) => msg.len(),
-            NotificationBuilder::ProcessDestroyed(msg) => msg.len(),
-        }
-    }
-
-    // TODO: change to a more strongly typed API
-    pub fn into_bytes(self) -> Vec<u8> {
-        match self {
-            NotificationBuilder::Interface(msg) => msg.into_bytes(),
-            NotificationBuilder::Response(msg) => msg.into_bytes(),
-            NotificationBuilder::ProcessDestroyed(msg) => msg.into_bytes(),
-        }
-    }
-
-    /// Modifies the `index_in_list` field of the message in construction.
-    pub fn set_index_in_list(&mut self, value: u32) {
-        match self {
-            NotificationBuilder::Interface(msg) => msg.set_index_in_list(value),
-            NotificationBuilder::Response(msg) => msg.set_index_in_list(value),
-            NotificationBuilder::ProcessDestroyed(msg) => msg.set_index_in_list(value),
-        }
-    }
-}
-
-impl From<InterfaceNotificationBuilder> for NotificationBuilder {
-    fn from(msg: InterfaceNotificationBuilder) -> NotificationBuilder {
-        NotificationBuilder::Interface(msg)
-    }
-}
-
-impl From<ResponseNotificationBuilder> for NotificationBuilder {
-    fn from(msg: ResponseNotificationBuilder) -> NotificationBuilder {
-        NotificationBuilder::Response(msg)
-    }
-}
-
-impl From<ProcessDestroyedNotificationBuilder> for NotificationBuilder {
-    fn from(msg: ProcessDestroyedNotificationBuilder) -> NotificationBuilder {
-        NotificationBuilder::ProcessDestroyed(msg)
-    }
-}
-
-/// Message received from the kernel.
-#[derive(Debug, Clone)]
-pub enum DecodedNotification {
-    /// Interface notification.
-    Interface(DecodedInterfaceNotification),
-    /// Response notification.
-    Response(DecodedResponseNotification),
-    /// Process destroyed notification.
-    ///
-    /// Whenever a process that has emitted events on one of our interfaces stops, a
-    /// `ProcessDestroyed` notification is sent.
-    ///
-    /// Note that this is done on a "best effort" basis. It is possible that these notifications
-    /// don't get immediately delivered, and that spurious notifications get emitted.
-    ProcessDestroyed(DecodedProcessDestroyedNotification),
-}
-
-// TODO: all the decoding performs unaligned reads, which isn't great
-
-/// Attempt to decode a notification.
-pub fn decode_notification(buffer: &[u8]) -> Result<DecodedNotification, ()> {
-    if buffer.is_empty() {
-        return Err(());
-    }
-
-    match buffer[0] {
-        0 => decode_interface_notification(buffer).map(DecodedNotification::Interface),
-        1 => decode_response_notification(buffer).map(DecodedNotification::Response),
-        2 => {
-            decode_process_destroyed_notification(buffer).map(DecodedNotification::ProcessDestroyed)
-        }
-        _ => Err(()),
-    }
-}
-
-/// Either a decoded interface notification or a decoded process destroyed notification.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DecodedInterfaceOrDestroyed {
-    /// Interface notification.
-    Interface(DecodedInterfaceNotification),
-    /// Process destroyed notification.
-    ProcessDestroyed(DecodedProcessDestroyedNotification),
-}
-
-/// Builds a interface notification from its raw components.
-pub fn build_interface_notification(
-    interface: &InterfaceHash,
-    message_id: Option<MessageId>,
-    emitter_pid: Pid,
-    index_in_list: u32,
-    actual_data: &EncodedMessage,
-) -> InterfaceNotificationBuilder {
-    let mut buffer = Vec::with_capacity(1 + 32 + 8 + 8 + 4 + actual_data.0.len());
-    buffer.push(0);
-    buffer.extend_from_slice(&interface.0);
-    buffer.extend_from_slice(&message_id.map(u64::from).unwrap_or(0).to_le_bytes());
-    buffer.extend_from_slice(&u64::from(emitter_pid).to_le_bytes());
-    buffer.extend_from_slice(&index_in_list.to_le_bytes());
-    buffer.extend_from_slice(&actual_data.0);
-
-    debug_assert_eq!(buffer.capacity(), buffer.len());
-    InterfaceNotificationBuilder { data: buffer }
-}
-
-#[derive(Debug, Clone)]
-pub struct InterfaceNotificationBuilder {
-    data: Vec<u8>,
-}
-
-impl InterfaceNotificationBuilder {
-    /// Updates the `index_in_list` field of the message.
-    pub fn set_index_in_list(&mut self, value: u32) {
-        self.data[49..53].copy_from_slice(&value.to_le_bytes());
-    }
-
-    /// Returns the [`MessageId`] that was put in the builder.
-    pub fn message_id(&self) -> Option<MessageId> {
-        let id = u64::from_le_bytes([
-            self.data[33],
-            self.data[34],
-            self.data[35],
-            self.data[36],
-            self.data[37],
-            self.data[38],
-            self.data[39],
-            self.data[40],
-        ]);
-
-        MessageId::try_from(id).ok()
-    }
-
-    pub fn len(&self) -> usize {
-        self.data.len()
-    }
-
-    pub fn into_bytes(self) -> Vec<u8> {
-        self.data
-    }
-}
-
-pub fn decode_interface_notification(buffer: &[u8]) -> Result<DecodedInterfaceNotification, ()> {
-    if buffer.len() < 1 + 32 + 8 + 8 + 4 {
-        return Err(());
-    }
-
-    if buffer[0] != 0x0 {
-        return Err(());
-    }
-
-    Ok(DecodedInterfaceNotification {
-        interface: InterfaceHash({
-            let mut hash = [0; 32];
-            hash.copy_from_slice(&buffer[1..33]);
-            hash
-        }),
-        message_id: {
-            let id = u64::from_le_bytes([
-                buffer[33], buffer[34], buffer[35], buffer[36], buffer[37], buffer[38], buffer[39],
-                buffer[40],
-            ]);
-
-            MessageId::try_from(id).ok()
-        },
-        emitter_pid: From::from(u64::from_le_bytes([
-            buffer[41], buffer[42], buffer[43], buffer[44], buffer[45], buffer[46], buffer[47],
-            buffer[48],
-        ])),
-        index_in_list: u32::from_le_bytes([buffer[49], buffer[50], buffer[51], buffer[52]]),
-        actual_data: EncodedMessage(buffer[53..].to_vec()),
-    })
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DecodedInterfaceNotification {
-    /// Interface the message concerns.
-    pub interface: InterfaceHash,
-    /// Id of the message. Can be used for answering. `None` if no answer is expected.
-    pub message_id: Option<MessageId>,
-    /// Id of the process that emitted the message.
-    ///
-    /// This should be used for security purposes, so that a process can't modify another process'
-    /// resources.
-    pub emitter_pid: Pid,
-    /// Index within the list to poll where this message was.
-    pub index_in_list: u32,
-    pub actual_data: EncodedMessage,
-}
-
-pub fn build_response_notification(
-    message_id: MessageId,
-    index_in_list: u32,
-    actual_data: Result<&EncodedMessage, ()>,
-) -> ResponseNotificationBuilder {
-    let mut buffer =
-        Vec::with_capacity(1 + 8 + 4 + 1 + actual_data.map(|m| m.0.len()).unwrap_or(0));
-    buffer.push(1);
-    buffer.extend_from_slice(&u64::from(message_id).to_le_bytes());
-    buffer.extend_from_slice(&index_in_list.to_le_bytes());
-    if let Ok(actual_data) = actual_data {
-        buffer.push(0);
-        buffer.extend_from_slice(&actual_data.0);
-    } else {
-        buffer.push(1);
-    }
-
-    debug_assert_eq!(buffer.capacity(), buffer.len());
-    ResponseNotificationBuilder { data: buffer }
-}
-
-#[derive(Debug, Clone)]
-pub struct ResponseNotificationBuilder {
-    data: Vec<u8>,
-}
-
-impl ResponseNotificationBuilder {
     /// Updates the `index_in_list` field of the message.
     pub fn set_index_in_list(&mut self, value: u32) {
         self.data[9..13].copy_from_slice(&value.to_le_bytes());
@@ -400,7 +174,7 @@ impl ResponseNotificationBuilder {
     }
 }
 
-pub fn decode_response_notification(buffer: &[u8]) -> Result<DecodedResponseNotification, ()> {
+pub fn decode_notification(buffer: &[u8]) -> Result<DecodedNotification, ()> {
     if buffer.len() < 1 + 8 + 4 + 1 {
         return Err(());
     }
@@ -414,7 +188,7 @@ pub fn decode_response_notification(buffer: &[u8]) -> Result<DecodedResponseNoti
         return Err(());
     }
 
-    Ok(DecodedResponseNotification {
+    Ok(DecodedNotification {
         message_id: MessageId::try_from({
             u64::from_le_bytes([
                 buffer[1], buffer[2], buffer[3], buffer[4], buffer[5], buffer[6], buffer[7],
@@ -432,7 +206,7 @@ pub fn decode_response_notification(buffer: &[u8]) -> Result<DecodedResponseNoti
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DecodedResponseNotification {
+pub struct DecodedNotification {
     /// Identifier of the message whose answer we are receiving.
     pub message_id: MessageId,
 
@@ -447,66 +221,6 @@ pub struct DecodedResponseNotification {
     pub actual_data: Result<EncodedMessage, ()>,
 }
 
-pub fn build_process_destroyed_notification(
-    pid: Pid,
-    index_in_list: u32,
-) -> ProcessDestroyedNotificationBuilder {
-    let mut buffer = Vec::with_capacity(1 + 8 + 4);
-    buffer.push(2);
-    buffer.extend_from_slice(&u64::from(pid).to_le_bytes());
-    buffer.extend_from_slice(&index_in_list.to_le_bytes());
-
-    debug_assert_eq!(buffer.capacity(), buffer.len());
-    ProcessDestroyedNotificationBuilder { data: buffer }
-}
-
-#[derive(Debug, Clone)]
-pub struct ProcessDestroyedNotificationBuilder {
-    data: Vec<u8>,
-}
-
-impl ProcessDestroyedNotificationBuilder {
-    /// Updates the `index_in_list` field of the message.
-    pub fn set_index_in_list(&mut self, value: u32) {
-        self.data[9..13].copy_from_slice(&value.to_le_bytes());
-    }
-
-    pub fn len(&self) -> usize {
-        self.data.len()
-    }
-
-    pub fn into_bytes(self) -> Vec<u8> {
-        self.data
-    }
-}
-
-pub fn decode_process_destroyed_notification(
-    buffer: &[u8],
-) -> Result<DecodedProcessDestroyedNotification, ()> {
-    if buffer.len() != 1 + 8 + 4 {
-        return Err(());
-    }
-
-    if buffer[0] != 0x2 {
-        return Err(());
-    }
-
-    Ok(DecodedProcessDestroyedNotification {
-        pid: From::from(u64::from_le_bytes([
-            buffer[1], buffer[2], buffer[3], buffer[4], buffer[5], buffer[6], buffer[7], buffer[8],
-        ])),
-        index_in_list: u32::from_le_bytes([buffer[9], buffer[10], buffer[11], buffer[12]]),
-    })
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DecodedProcessDestroyedNotification {
-    /// Identifier of the process that got destroyed.
-    pub pid: Pid,
-    /// Index within the list to poll where this message was.
-    pub index_in_list: u32,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -514,67 +228,18 @@ mod tests {
     use core::{convert::TryFrom, num::NonZeroU64};
 
     #[test]
-    fn interface_message_encode_decode() {
-        let interface_hash = From::from([0xca; 32]);
-        let message_id = Some(TryFrom::try_from(0x0123456789abcdef).unwrap());
-        let pid = From::from(0xfedcba9876543210);
-        let index_in_list = 0xdeadbeef;
-        let message = EncodedMessage(vec![8, 7, 9]);
-
-        let mut int_notif =
-            build_interface_notification(&interface_hash, message_id, pid, 0xf00baa, &message);
-        int_notif.set_index_in_list(index_in_list);
-        assert_eq!(int_notif.message_id(), message_id);
-
-        let decoded = decode_interface_notification(&int_notif.into_bytes()).unwrap();
-        assert_eq!(decoded.interface, interface_hash);
-        assert_eq!(decoded.message_id, message_id);
-        assert_eq!(decoded.emitter_pid, pid);
-        assert_eq!(decoded.index_in_list, index_in_list);
-        assert_eq!(decoded.actual_data, message);
-    }
-
-    #[test]
     fn response_message_encode_decode() {
         let message_id = TryFrom::try_from(0x0123456789abcdef).unwrap();
         let index_in_list = 0xdeadbeef;
         let message = EncodedMessage(vec![8, 7, 9]);
 
-        let mut resp_notif = build_response_notification(message_id, 0xf00baa, Ok(&message));
+        let mut resp_notif = build_notification(message_id, 0xf00baa, Ok(&message));
         resp_notif.set_index_in_list(index_in_list);
         assert_eq!(resp_notif.message_id(), message_id);
 
-        let decoded = decode_response_notification(&resp_notif.into_bytes()).unwrap();
+        let decoded = decode_notification(&resp_notif.into_bytes()).unwrap();
         assert_eq!(decoded.message_id, message_id);
         assert_eq!(decoded.index_in_list, index_in_list);
         assert_eq!(decoded.actual_data, Ok(message));
-    }
-
-    #[test]
-    fn response_message_err_encode_decode() {
-        let message_id = TryFrom::try_from(0x0123456789abcdef).unwrap();
-        let index_in_list = 0xdeadbeef;
-
-        let mut resp_notif = build_response_notification(message_id, 0xf00baa, Err(()));
-        resp_notif.set_index_in_list(index_in_list);
-        assert_eq!(resp_notif.message_id(), message_id);
-
-        let decoded = decode_response_notification(&resp_notif.into_bytes()).unwrap();
-        assert_eq!(decoded.message_id, message_id);
-        assert_eq!(decoded.index_in_list, index_in_list);
-        assert_eq!(decoded.actual_data, Err(()));
-    }
-
-    #[test]
-    fn process_destroyed_message_encode_decode() {
-        let pid = From::from(0xfedcba9876543210);
-        let index_in_list = 0xdeadbeef;
-
-        let mut destr_notif = build_process_destroyed_notification(pid, 0xf00baa);
-        destr_notif.set_index_in_list(index_in_list);
-
-        let decoded = decode_process_destroyed_notification(&destr_notif.into_bytes()).unwrap();
-        assert_eq!(decoded.pid, pid);
-        assert_eq!(decoded.index_in_list, index_in_list);
     }
 }
